+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT-RELEASE
+
+# v0.5.0
 - Added `get_default(name)` function to scripting. It allows lookup of constants like the `Shopware/src/Core/Defaults.php`
 - Fixed import of "To-Many-Associations" when the value is null it will be ignored instead of added to the entity